+++ docs/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update field `updated_dt` on queryset update (OSIDB-3573)
 - Introduce purl to Affect (OSIDB-3409)
 
+### Changed
+- Add history to several other models: AffectCVSS, FlawAcknowledgment, FlawComment,
+  FlawCVSS, FlawReference, Snippet and Tracker. (OSIDB-3466)
+
 ## [4.5.6] - 2024-11-08
 ### Fixed
 - Properly save updated flaw in NVD collector (OSIDB-3661)